finished recipes and ingredients creation process

+++ src/main/java/com/cookplanner/repositories/Roles.java
@@ -13,7 +13,7 @@ import java.util.List;
 
 
 @Repository
-public interface Roles extends CrudRepository<UserRole, Long> {
+public interface Roles extends CrudRepository<UserRole, Integer> {
     @Query("select ur.role from UserRole ur, User u where u.username=?1 and ur.userId = u.id")
     public List<String> ofUserWith (String username);
