+++ src/main/java/com/cookplanner/repositories/Users.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
  */
 
 @Repository
-public interface Users extends CrudRepository<User, Long> {
+public interface Users extends CrudRepository<User, Integer> {
 
     public User findByUsername(String username);
 }
